requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
linkedin-api>=2.0.0
selenium>=4.16.0
fastapi>=0.104.0
pydantic>=2.4.0
python-jose>=3.3.0
pytest>=7.4.0
aiosqlite>=0.19.0
alembic>=1.12.0
uvicorn>=0.24.0
python-multipart>=0.0.6
passlib>=1.7.4
bcrypt>=4.0.1
pydantic-settings>=2.0.0
email-validator>=2.0.0
redis>=5.0.1
structlog>=24.1.0
webdriver_manager>=4.0.1
